Rename side menu and fixed layout
