+++ react/components/SideMenu/SideMenu.js
