+++ react/editor/style.js
@@ -3,7 +3,7 @@ import styled from "styled-components";
 const Container = styled.div`
   display: flex;
   width: 100%;
-  height: 100%;
+  height: calc(100vh - 47px);
 `;
 
 const EditSection = styled.div`
